model_v45/requirements.txt
# Python зависимости для модели "nauslava/rubber-tile-lora-v45"
# Версии оптимизированы для стабильности и производительности

# Основные ML библиотеки
torch>=2.0.0,<3.0.0
diffusers>=0.27.0,<0.28.0
transformers>=4.47.0,<4.48.0
accelerate>=0.27.0,<0.28.0

# Безопасность и производительность
safetensors>=0.4.0,<0.5.0

# Обработка изображений
Pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0

# Оптимизации
xformers>=0.0.22,<0.1.0
optimum>=1.16.0,<1.17.0

# Дополнительные утилиты
requests>=2.31.0,<3.0.0
tqdm>=4.65.0,<5.0.0

# Безопасность
cryptography>=41.0.0,<42.0.0
